pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sc-smart-device"
version = "3.2.0"
description = "Provides a simple, consistent access to Shelly and Tasmota smart switches and energy meters."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sc-foundation-services>=3.0.3",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-mock>=3.15.1",
    "pre-commit>=3.5.0",
    "pytest-dotenv>=0.5.2",
]
docs = [
    "mkdocs>=1.6.1,<2.0.0",
    "mkdocs-include-markdown-plugin>=6.2.2",
    "mkdocs-material>=9.6.14",
    "mkdocstrings>=0.26.1",
    "mkdocstrings-python>=1.11.1",
    "pdoc>=14.7.0",
]
all = [
    "sc-smart-device[dev,docs]",
]

[project.urls]
Homepage = "https://github.com/NickElseySpelloC"
Repository = "https://github.com/NickElseySpelloC/sc-smart-device"

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true
zip-safe = false

[tool.setuptools.package-data]
sc_smart_device = ["shelly_models.json"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["tests", "."]

[dev.env.sync]
remote="~/Library/CloudStorage/Dropbox/Development/dev_setup/app_config/sc-smart-device"
patterns = [
    ".env",
    ".vscode/*.json",
    "development/*.yaml",
    "development/*.log",
    "tests/config.yaml"
]
